requirements.txt
awscrt==0.19.1
awsiotsdk==1.19.0
boto3==1.28.62
botocore==1.31.62
jmespath==1.0.1
python-dateutil==2.8.2
s3transfer==0.7.0
six==1.16.0
urllib3==1.26.17
